+++ examples/chain_client/0_LocalOrderHash.py
@@ -2,6 +2,7 @@ import asyncio
 import uuid
 
 from pyinjective.async_client import AsyncClient
+from pyinjective.constant import GAS_FEE_BUFFER_AMOUNT, GAS_PRICE
 from pyinjective.core.network import Network
 from pyinjective.orderhash import OrderHashManager
 from pyinjective.transaction import Transaction
@@ -101,9 +102,9 @@ async def main() -> None:
         .with_account_num(client.get_number())
         .with_chain_id(network.chain_id)
     )
-    gas_price = 500000000
+    gas_price = GAS_PRICE
     base_gas = 85000
-    gas_limit = base_gas + 20000  # add 20k for gas, fee computation
+    gas_limit = base_gas + GAS_FEE_BUFFER_AMOUNT  # add buffer for gas fee computation
     gas_fee = "{:.18f}".format((gas_price * gas_limit) / pow(10, 18)).rstrip("0")
     fee = [
         composer.Coin(
@@ -138,9 +139,9 @@ async def main() -> None:
         .with_account_num(client.get_number())
         .with_chain_id(network.chain_id)
     )
-    gas_price = 500000000
+    gas_price = GAS_PRICE
     base_gas = 85000
-    gas_limit = base_gas + 20000  # add 20k for gas, fee computation
+    gas_limit = base_gas + GAS_FEE_BUFFER_AMOUNT  # add buffer for gas fee computation
     gas_fee = "{:.18f}".format((gas_price * gas_limit) / pow(10, 18)).rstrip("0")
     fee = [
         composer.Coin(
@@ -228,9 +229,9 @@ async def main() -> None:
         .with_account_num(client.get_number())
         .with_chain_id(network.chain_id)
     )
-    gas_price = 500000000
+    gas_price = GAS_PRICE
     base_gas = 85000
-    gas_limit = base_gas + 20000  # add 20k for gas, fee computation
+    gas_limit = base_gas + GAS_FEE_BUFFER_AMOUNT  # add buffer for gas fee computation
     gas_fee = "{:.18f}".format((gas_price * gas_limit) / pow(10, 18)).rstrip("0")
     fee = [
         composer.Coin(